refactor(settings-v2): extract ProviderDataBlock + add Suspense fallback (#1450)
- Replace null Suspense fallback in SettingsV2Content with SectionTitle so the
  header anchors during lazy-chunk hydration (no body flicker on slow connections)
- Extract ProviderDataBlock sub-component to sections/ProviderDataBlock.tsx
- Extract styled-component cluster to sections/AdvancedSection.styled.ts
- AdvancedSection.tsx reduced from 511 to 260 lines (well under 500-line guideline)
- Add TODO(#1462) in About block for deferred build-time version constant
