+++ src/components/SettingsV2/sections/AdvancedSection.tsx
@@ -1,26 +1,35 @@
-import React, { useCallback, useMemo, useRef, useState } from 'react';
-import styled from 'styled-components';
+import React, { useCallback, useMemo, useState } from 'react';
 
 import { useProviderContext } from '@/contexts/ProviderContext';
 import { useProfilingContext } from '@/contexts/ProfilingContext';
 import { useVisualizerDebug } from '@/contexts/VisualizerDebugContext';
 import { useQapEnabled } from '@/hooks/useQapEnabled';
-import { useAsyncAction } from '@/hooks/useAsyncAction';
-import { ART_REFRESHED_EVENT } from '@/hooks/useLibrarySync';
 import { clearCacheWithOptions } from '@/services/cache/libraryCache';
 import { clearAllPins } from '@/services/settings/pinnedItemsStorage';
 import { STORAGE_KEYS } from '@/constants/storage';
 import { STATUS_RESET_DELAY_MS } from '@/constants/statusTiming';
-import type { CatalogProvider } from '@/types/providers';
 
 import { Switch } from '@/components/ui/switch';
 import { Button } from '@/components/ui/button';
+import { Accordion } from '@/components/ui/accordion';
+
+import { ProviderDataBlock } from './ProviderDataBlock';
 import {
-  Accordion,
-  AccordionItem,
-  AccordionTrigger,
-  AccordionContent,
-} from '@/components/ui/accordion';
+  Container,
+  ControlBlock,
+  ControlRow,
+  ControlLabelText,
+  ControlHelp,
+  SectionGroupTitle,
+  Divider,
+  CacheOptionList,
+  CacheOptionItem,
+  CacheCheckbox,
+  CacheButtons,
+  ShortcutHintList,
+  ShortcutHintRow,
+  Kbd,
+} from './AdvancedSection.styled';
 
 /**
  * SettingsV2 Advanced section — phase 2 port (#1450).
@@ -37,267 +46,6 @@ import {
  *   - Clear Library Cache → `clearCacheWithOptions()` + `clearAllPins()` +
  *     accent-overrides cleanup (mirrors `PlayerControlsSection.handleClearCache`).
  */
-
-const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: ${({ theme }) => theme.spacing.xl};
-`;
-
-const ControlBlock = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: ${({ theme }) => theme.spacing.sm};
-`;
-
-const ControlRow = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: ${({ theme }) => theme.spacing.md};
-`;
-
-const ControlLabelText = styled.label`
-  font-size: ${({ theme }) => theme.fontSize.sm};
-  color: hsl(var(--foreground));
-  font-weight: ${({ theme }) => theme.fontWeight.medium};
-`;
-
-const ControlHelp = styled.p`
-  margin: 0;
-  font-size: ${({ theme }) => theme.fontSize.xs};
-  color: hsl(var(--muted-foreground));
-`;
-
-const SectionGroupTitle = styled.h4`
-  margin: 0 0 ${({ theme }) => theme.spacing.sm};
-  font-size: ${({ theme }) => theme.fontSize.sm};
-  font-weight: ${({ theme }) => theme.fontWeight.semibold};
-  color: hsl(var(--foreground));
-  text-transform: uppercase;
-  letter-spacing: 0.05em;
-`;
-
-const Divider = styled.div`
-  border-top: 1px solid hsl(var(--border));
-`;
-
-const CacheOptionList = styled.ul`
-  list-style: none;
-  margin: 0;
-  padding: 0;
-  display: flex;
-  flex-direction: column;
-  gap: ${({ theme }) => theme.spacing.sm};
-`;
-
-const CacheOptionItem = styled.li`
-  display: flex;
-  align-items: center;
-  gap: ${({ theme }) => theme.spacing.sm};
-`;
-
-const CacheCheckbox = styled.input`
-  appearance: none;
-  -webkit-appearance: none;
-  width: 16px;
-  height: 16px;
-  border: 1px solid hsl(var(--border));
-  border-radius: ${({ theme }) => theme.borderRadius.sm};
-  background: hsl(var(--input));
-  cursor: pointer;
-  flex-shrink: 0;
-  position: relative;
-  transition: all ${({ theme }) => theme.transitions.fast};
-
-  &:checked {
-    background: hsl(var(--primary));
-    border-color: hsl(var(--primary));
-  }
-
-  &:checked::after {
-    content: '';
-    position: absolute;
-    left: 4px;
-    top: 1px;
-    width: 5px;
-    height: 9px;
-    border: 2px solid hsl(var(--primary-foreground));
-    border-top: none;
-    border-left: none;
-    transform: rotate(45deg);
-  }
-
-  &:focus-visible {
-    outline: 2px solid hsl(var(--ring));
-    outline-offset: 2px;
-  }
-`;
-
-const CacheButtons = styled.div`
-  display: flex;
-  gap: ${({ theme }) => theme.spacing.sm};
-`;
-
-const ShortcutHintList = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: ${({ theme }) => theme.spacing.xs};
-`;
-
-const ShortcutHintRow = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  font-size: ${({ theme }) => theme.fontSize.xs};
-  color: hsl(var(--muted-foreground));
-`;
-
-const Kbd = styled.kbd`
-  background: hsl(var(--muted));
-  border: 1px solid hsl(var(--border));
-  border-radius: ${({ theme }) => theme.borderRadius.sm};
-  padding: 0 ${({ theme }) => theme.spacing.xs};
-  font-family: 'Monaco', 'Menlo', monospace;
-  font-size: ${({ theme }) => theme.fontSize.xs};
-  color: hsl(var(--foreground));
-`;
-
-interface ProviderDataBlockProps {
-  providerName: string;
-  catalog: CatalogProvider;
-}
-
-const ProviderDataBlock: React.FC<ProviderDataBlockProps> = ({ providerName, catalog }) => {
-  const [resultMessage, setResultMessage] = useState('');
-  const fileInputRef = useRef<HTMLInputElement>(null);
-
-  const hasArtCache = !!catalog.clearArtCache;
-  const hasRefreshArt = !!catalog.refreshArtCache;
-  const hasLikesManagement = !!catalog.exportLikes && !!catalog.importLikes;
-  const hasMetadataRefresh = !!catalog.refreshLikedMetadata;
-
-  const clearResultMessage = useCallback(() => setResultMessage(''), []);
-
-  const clearArtFn = useCallback(async () => {
-    await catalog.clearArtCache?.();
-  }, [catalog]);
-
-  const refreshArtFn = useCallback(async () => {
-    await catalog.refreshArtCache?.();
-    window.dispatchEvent(new CustomEvent(ART_REFRESHED_EVENT));
-  }, [catalog]);
-
-  const exportFn = useCallback(async () => {
-    try {
-      const json = await catalog.exportLikes!();
-      const blob = new Blob([json], { type: 'application/json' });
-      const url = URL.createObjectURL(blob);
-      const a = document.createElement('a');
-      a.href = url;
-      a.download = `vorbis-liked-songs-${new Date().toISOString().slice(0, 10)}.json`;
-      a.click();
-      URL.revokeObjectURL(url);
-      setResultMessage('Exported!');
-    } catch {
-      setResultMessage('Export failed');
-    }
-  }, [catalog]);
-
-  const refreshMetadataFn = useCallback(async () => {
-    try {
-      const result = await catalog.refreshLikedMetadata!();
-      const parts: string[] = [];
-      if (result.updated > 0) parts.push(`${result.updated} updated`);
-      if (result.removed > 0) parts.push(`${result.removed} removed`);
-      setResultMessage(parts.length > 0 ? parts.join(', ') : 'No changes');
-    } catch {
-      setResultMessage('Refresh failed');
-    }
-  }, [catalog]);
-
-  const [clearArtStatus, runClearArt] = useAsyncAction(clearArtFn);
-  const [refreshArtStatus, runRefreshArt] = useAsyncAction(refreshArtFn);
-  const [exportStatus, runExport] = useAsyncAction(exportFn, { onReset: clearResultMessage });
-  const [metadataStatus, runRefreshMetadata] = useAsyncAction(refreshMetadataFn, { onReset: clearResultMessage });
-  const [importStatus, setImportStatus] = useState<'idle' | 'working' | 'done'>('idle');
-
-  const handleImport = useCallback(async (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    setImportStatus('working');
-    try {
-      const json = await file.text();
-      const count = await catalog.importLikes!(json);
-      setResultMessage(`Imported ${count} tracks`);
-    } catch {
-      setResultMessage('Import failed');
-    }
-    if (fileInputRef.current) fileInputRef.current.value = '';
-    setImportStatus('done');
-    setTimeout(() => { setImportStatus('idle'); setResultMessage(''); }, STATUS_RESET_DELAY_MS);
-  }, [catalog]);
-
-  const artBusy = clearArtStatus === 'working' || refreshArtStatus === 'working';
-  const likesBusy = exportStatus === 'working' || importStatus === 'working' || metadataStatus === 'working';
-
-  return (
-    <AccordionItem value={`${providerName.toLowerCase()}-data`}>
-      <AccordionTrigger>{`${providerName} Data`}</AccordionTrigger>
-      <AccordionContent>
-        <ControlBlock>
-          {hasArtCache && (
-            <ControlRow>
-              <ControlHelp>Clear cached art so it re-downloads on next library load</ControlHelp>
-              <Button variant="outline" size="sm" onClick={runClearArt} disabled={artBusy}>
-                {clearArtStatus === 'done' ? 'Cleared!' : artBusy ? 'Working…' : 'Clear Art Cache'}
-              </Button>
-            </ControlRow>
-          )}
-          {hasRefreshArt && (
-            <ControlRow>
-              <ControlHelp>Clear and immediately re-fetch fresh art in the background</ControlHelp>
-              <Button variant="outline" size="sm" onClick={runRefreshArt} disabled={artBusy}>
-                {refreshArtStatus === 'done' ? 'Started!' : artBusy ? 'Working…' : 'Refresh Art'}
-              </Button>
-            </ControlRow>
-          )}
-          {hasLikesManagement && (
-            <>
-              <ControlRow>
-                <ControlHelp>Export liked songs to a JSON file for backup</ControlHelp>
-                <Button variant="outline" size="sm" onClick={runExport} disabled={likesBusy}>
-                  {exportStatus === 'done' && resultMessage === 'Exported!' ? 'Exported!' : likesBusy ? 'Working…' : 'Export Likes'}
-                </Button>
-              </ControlRow>
-              <ControlRow>
-                <ControlHelp>Import liked songs from a previously exported JSON file</ControlHelp>
-                <input
-                  ref={fileInputRef}
-                  type="file"
-                  accept=".json"
-                  onChange={handleImport}
-                  style={{ display: 'none' }}
-                />
-                <Button variant="outline" size="sm" onClick={() => fileInputRef.current?.click()} disabled={likesBusy}>
-                  {importStatus === 'done' && resultMessage.startsWith('Imported') ? resultMessage : likesBusy ? 'Working…' : 'Import Likes'}
-                </Button>
-              </ControlRow>
-            </>
-          )}
-          {hasMetadataRefresh && (
-            <ControlRow>
-              <ControlHelp>Re-scan {providerName} to update metadata for liked tracks</ControlHelp>
-              <Button variant="outline" size="sm" onClick={runRefreshMetadata} disabled={likesBusy}>
-                {metadataStatus === 'done' ? resultMessage || 'Done!' : likesBusy ? 'Scanning…' : 'Refresh Metadata'}
-              </Button>
-            </ControlRow>
-          )}
-        </ControlBlock>
-      </AccordionContent>
-    </AccordionItem>
-  );
-};
 
 export const AdvancedSection: React.FC = () => {
   const { enabledProviderIds, registry } = useProviderContext();
@@ -490,6 +238,7 @@ export const AdvancedSection: React.FC = () => {
 
       <ControlBlock>
         <SectionGroupTitle>About</SectionGroupTitle>
+        {/* TODO(#1462): surface app version once a build-time constant is wired */}
         <ControlHelp>Vorbis Player — keyboard-first music player.</ControlHelp>
         <ShortcutHintList>
           <ShortcutHintRow>
